Organize SSV parsing utilities into data_parser module
