+++ README.md
@@ -15,8 +15,9 @@ This project opens the BGF Retail store login page using Selenium. It is a simpl
 
 The mid-category sales automation is now executed directly from `main.py` using
 `modules/sales_analysis/mid_category_sales_ssv.json` to capture the dataset.
-After downloading the SSV response, `parse_and_save.py` parses the data and
-stores rows with zero stock quantity in `output/category_001_filtered.txt`.
+After downloading the SSV response, `modules/data_parser/parse_and_save.py`
+parses the data and stores rows with zero stock quantity in
+`output/category_001_filtered.txt`.
 
 
 The structure files in the `structure` directory describe the XPath selectors
